requirements.txt
awscli==1.29.59
black==23.9.1
boto3==1.28.59
botocore==1.31.59
click==8.1.7
colorama==0.4.4
docutils==0.16
EasyProcess==1.1
entrypoint2==1.1
jmespath==1.0.1
MouseInfo==0.1.3
mss==9.0.1
mypy-extensions==1.0.0
numpy==1.24.4
opencv-python==4.8.1.78
packaging==23.2
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.11.0
pyasn1==0.5.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySimpleGUI==4.60.5
python-dateutil==2.8.2
pytweening==1.0.7
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.7.0
six==1.16.0
tomli==2.0.1
typing-extensions==4.8.0
urllib3==1.26.17
